Refactor: Modularize large files and reduce line counts under 200 lines
- Split schemas.py into sub-modules in schemas/ package
- Modularize models/__init__.py into domain-specific files
- Extract schema initialization from database.py to infrastructure/schema.py
- Separate world state management from SessionRepository to WorldStateRepository
- Split router_TRACE.py into Turn and Phase history routers
- Remove redundant and legacy SQL/Python files
- Fix path priority in TRACE routers
- All 71 tests passed
