+++ docs/current_task.md
@@ -1,25 +1,12 @@
-# Current Task: Post-merge System Integrity Verification
+# Current Task: Secondary Python Refactoring
 
-## 목적
+## 목표
 
-- 다른 작업자의 변경 사항과 머지된 후 시스템의 정합성 및 기능적 무결성 검증
-- 전체 테스트 스위트를 실행하여 회귀 오류 발생 여부 확인
+200줄 이상의 파이썬 파일을 리팩터링하여 가독성과 유지보수성을 극대화함.
 
-## 현황
+## 작업 내용
 
-- 머지 완료 직후 상태
-- `CORE_ENGINE_HANDBOOK.md` 내용 확인 완료
-- 주요 데이터 아키텍처(Session 0 Deep Copy 등) 및 기술 스택 파악 완료
-
-## 계획
-
-1. `uv run pytest tests/ -v` 명령어를 통해 전체 테스트 실행
-2. 실패하는 테스트가 있을 경우 로그 분석 및 원인 파악
-3. 테스트 결과 요약 및 업데이트
-
-## 진행 상황
-
-- [x] 전체 테스트 실행
-- [x] 순환 참조(Circular Import) 오류 수정 (`configs/__init__.py`)
-- [x] 레포지토리 내 SQL 파일 경로 불일치 수정 (`entity.py`, `player.py`, `session.py`)
-- [x] 전체 테스트 재실행 및 100% 통과 확인
+1. `database.py`의 초기화 로직 분리.
+2. `session.py` 리포지토리의 책임 분산.
+3. `router_TRACE.py` 분리.
+4. `models/__init__.py` 분할.
