+++ docs/plan.md
@@ -1,57 +1,66 @@
-📜 TRPG 시스템 데이터베이스 구축 계획서 (Master Plan)
+# Refactoring Plan
 
-1. 파일 구조 및 명명 규칙
+## 목적
 
-시스템은 정적 구조(Base)와 동적 로직(Logic)을 엄격히 분리하여 관리합니다.
+500줄 이상의 대형 파일들을 기능별로 분할하여 유지보수수를 높이고 가독성을 개선함.
 
-- B_ (Base): 테이블 스키마, 인덱스, 기본 시간 갱신 트리거를 포함합니다.
-- L_ (Logic): 세션 생성 시 Session 0 데이터를 복제하는 함수 및 초기화 트리거를 포함합니다.
+## 대상 파일
 
-1. 핵심 운영 원칙: Session 0 Deep Copy
+1. `src/state_db/schemas.py` (788 lines)
+2. `src/state_db/data/node/core/session_Query.sql` (766 lines)
+3. `src/state_db/Query/TRACE/phase_tracing-r.sql` (701 lines)
+4. `src/state_db/data/node/core/m_act_seq.sql` (692 lines)
+5. `src/state_db/Query/DEBUG/concept-r.sql` (647 lines)
 
-- Session 0: 00000000-0000-0000-0000-000000000000 ID를 시스템 마스터 세션으로 정의합니다.
-- Master Data: 작가가 작성한 시나리오 원본 데이터는 각 테이블에 session_id = '0...0'으로 저장됩니다.
-- Auto-Instantiation: 신규 세션 INSERT 시, 트리거가 마스터 데이터를 조회하여 해당 세션 ID로 복사본을 자동 생성합니다.
+## 진행 현황
 
-1. 상세 구축 로드맵
+- [x] `src/state_db/schemas.py` 분할
+- [x] `src/state_db/data/node/core/session_Query.sql` 분할 및 삭제 (중복 제거)
+- [x] `src/state_db/Query/TRACE/phase_tracing-r.sql` 분할
+- [x] `src/state_db/data/node/core/m_act_seq.sql` 분할 및 삭제 (중복 제거)
+- [x] `src/state_db/Query/DEBUG/concept-r.sql` 분할 및 삭제 (중복 제거)
 
-Phase 1: 기반 구조 구축 (Base)
+## 2차 리팩터링 (Python 파일 200줄 이하 목표)
 
-- B_scenario.sql & B_session.sql 실행
-  - 모든 엔티티의 부모가 되는 시나리오와 세션 테이블을 먼저 생성합니다.
-  - B_session.sql 내에서 시스템 세션(UUID 0) 레코드를 강제 삽입합니다.
-- 엔티티 테이블 생성 (B_npc.sql, B_enemy.sql, B_item.sql 등)
-  - 사용자 원본 구조를 유지하며 테이블만 생성합니다. (초기화 함수 미포함)
+- [ ] `src/state_db/infrastructure/database.py`: 스키마 초기화 로직 분리
+- [ ] `src/state_db/repositories/session.py`: 월드 상태 관리 로직 분리
+- [ ] `src/state_db/routers/router_TRACE.py`: Turn/Phase 라우터 분리
+- [ ] `src/state_db/models/__init__.py`: 모델 정의 파일 분산
 
-Phase 2: 복제 로직 주입 (Logic)
+## 상세 계획
 
-- 엔티티별 복제 함수 (L_*.sql) 등록
-  - WHERE session_id = '000...0' AND scenario_id = NEW.scenario_id 로직을 가진 함수를 등록합니다.
-- 세션 트리거 연결
-  - session 테이블에 AFTER INSERT 트리거를 연결하여 세션 생성과 동시에 모든 데이터가 복제되도록 설정합니다.
-  - 무한 루프 방지: WHEN (NEW.session_id <> '0...0') 조건을 반드시 적용합니다.
+### 1. `src/state_db/schemas.py` 분할
 
-Phase 3: 데이터 수급 및 검증 - [완료]
+- `src/state_db/schemas/` 디렉토리 생성
+- `base.py`: 공통 Enum 및 Base 클래스
+- `session.py`: 세션 관련 스키마
+- `inventory.py`: 인벤토리 관련 스키마
+- `item.py`: 아이템 관련 스키마
+- `player.py`: 플레이어 관련 스키마
+- `npc.py`: NPC 관련 스키마
+- `enemy.py`: 적(Enemy) 관련 스키마
+- `world.py`: 위치, Phase, Act, Sequence 관련 스키마
+- `scenario.py`: 시나리오 주입 관련 스키마
+- `__init__.py`: 모든 스키마 export 및 기존 경로 호환성 유지
 
-- [x] 마스터 데이터 로드: 시나리오 원형 데이터를 session_id = '0...0'으로 입력 완료.
-- [x] 기본 엔티티 복제 검증: 새 세션 생성 시 NPC, Enemy 자동 생성 확인.
-- [x] 관계 데이터 복제 검증 (RDB): player_npc_relations 테이블 초기화 확인.
-- [x] 관계 데이터 복제 검증 (Graph): Apache AGE 내 RELATION Edge 복제 확인.
+### 2. SQL 파일 분할
 
-1. 시나리오 주입 API (Scenario Injection API) - [완료]
+- SQL 파일들도 기능별로 쪼개어 가독성 확보
+- Apache AGE 쿼리 특성상 연관된 로직끼리 묶어서 분리
 
-목적: 작가나 관리자가 시나리오 메타데이터와 마스터 엔티티(NPC, Enemy 등)를 한 번의 API 호출로 시스템에 등록할 수 있도록 합니다.
+#### 2.1 `src/state_db/Query/TRACE/phase_tracing-r.sql` 분할
 
-기능:
+- `src/state_db/Query/TRACE/phase/` 디렉토리 하위로 이동
+- `get_history.sql`: 전체 Phase 전환 이력
+- `get_recent.sql`: 최근 Phase 전환
+- `get_by_phase.sql`: 특정 Phase 필터링
+- `get_range.sql`: Turn 범위 필터링
+- `get_latest.sql`: 최신 Phase 전환
+- `get_statistics.sql`: Phase 통계
+- `get_pattern.sql`: Phase 패턴
+- `get_summary.sql`: Phase 요약
 
-- [x] 시나리오 기본 정보 등록 (Title, Description 등)
-- [x] 마스터 NPC 데이터 등록 (Session 0)
-- [x] 마스터 Enemy 데이터 등록 (Session 0)
-- [x] 시나리오 관계 데이터(Graph) 등록 (Session 0)
-- [x] 테스트 코드 작성 및 검증 (test_logic_integration.py)
+#### 2.2 `src/state_db/data/node/core/session_Query.sql` 분할
 
-1. 시스템 안정화 및 통합 (2026-01-29) - [진행 중]
-
-- [x] 머지 후 순환 참조 오류 해결 및 구조 정적화.
-- [x] 디렉토리 구조 변경에 따른 SQL 쿼리 경로 전수 조사 및 반영.
-- [x] 전체 통합 테스트(71개 항목) 100% 통과 및 정합성 검증 완료.
+- `src/state_db/Query/INQUIRY/session/` 및 `src/state_db/Query/MANAGE/session/` 등으로 분산 배치
+- 중복되거나 통합된 쿼리들을 개별 파일로 분리하여 리포지토리에서 관리 용이하게 변경
