+++ src/state_db/Query/TRACE/phase/get_history.sql
@@ -1,0 +1,11 @@
+SELECT
+    history_id,
+    session_id,
+    previous_phase,
+    new_phase,
+    turn_at_transition,
+    transition_reason,
+    transitioned_at
+FROM phase_history
+WHERE session_id = $1
+ORDER BY transitioned_at ASC;
