+++ src/state_db/Query/TRACE/phase/get_statistics.sql
@@ -1,0 +1,1 @@
+SELECT * FROM get_phase_statistics($1);
