+++ src/state_db/infrastructure/database.py
@@ -183,6 +183,8 @@ async def run_cypher_query(
 
 async def startup() -> None:
     """애플리케이션 시작 시 초기화"""
+    from .schema import initialize_schema
+
     await DatabaseManager.get_pool()
 
     # 쿼리 로드 (상위 폴더의 Query 디렉토리)
@@ -193,68 +195,7 @@ async def startup() -> None:
     await init_age_graph()
 
     # 2. 스키마 초기화 (테이블 및 트리거 생성)
-    # B_ (Base) -> L_ (Logic) 순서로 실행
-    base_dir = query_dir / "BASE"
-
-    # 의존성을 고려한 실행 순서 정의
-    initial_tables = ["B_scenario.sql", "B_session.sql"]
-    entity_tables = [
-        "B_player.sql",
-        "B_npc.sql",
-        "B_enemy.sql",
-        "B_item.sql",
-        "B_phase.sql",
-        "B_turn.sql",
-    ]
-    relation_tables = [
-        "B_player_inventory.sql",
-        "B_player_npc_relations.sql",
-        "B_inventory.sql",
-    ]
-
-    async with DatabaseManager.get_connection() as conn:
-        await set_age_path(conn)
-
-        async with conn.transaction():
-            # 1단계: 기본 테이블
-            for filename in initial_tables:
-                file_path = base_dir / filename
-                if file_path.exists():
-                    print(f"Executing Stage 1 (Initial): {filename}")
-                    with open(file_path, "r", encoding="utf-8") as f:
-                        await conn.execute(f.read())
-
-            # 2단계: 주요 엔티티 테이블
-            for filename in entity_tables:
-                file_path = base_dir / filename
-                if file_path.exists():
-                    print(f"Executing Stage 2 (Entity): {filename}")
-                    with open(file_path, "r", encoding="utf-8") as f:
-                        await conn.execute(f.read())
-
-            # 3단계: 관계 및 기타 테이블
-            for filename in relation_tables:
-                file_path = base_dir / filename
-                if file_path.exists():
-                    print(f"Executing Stage 3 (Relation): {filename}")
-                    with open(file_path, "r", encoding="utf-8") as f:
-                        await conn.execute(f.read())
-
-            # 나머지 Base 파일들 (미처리된 것)
-            all_processed = set(initial_tables + entity_tables + relation_tables)
-            for file_path in base_dir.glob("B_*.sql"):
-                if file_path.name not in all_processed:
-                    print(f"Executing Stage 4 (Other Base): {file_path.name}")
-                    with open(file_path, "r", encoding="utf-8") as f:
-                        await conn.execute(f.read())
-
-            # 5단계: 로직 파일 (Triggers/Functions)
-            for file_path in sorted(base_dir.glob("L_*.sql")):
-                print(f"Executing Stage 5 (Logic): {file_path.name}")
-                with open(file_path, "r", encoding="utf-8") as f:
-                    await conn.execute(f.read())
-
-    logger.info("✅ Database schema and logic initialization completed.")
+    await initialize_schema(query_dir)
 
 
 async def shutdown() -> None:
