+++ src/state_db/schemas/base.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class Phase(str, Enum):
+    """게임 진행 단계 (Phase)"""
+
+    EXPLORATION = "exploration"
+    COMBAT = "combat"
+    DIALOGUE = "dialogue"
+    REST = "rest"
